Fix overwriting prev_move_pos
When creating children for a node, prev_move_pos was being overwritten
when it shouldn't have been. Instead, we just pass in the pos to a
child node when it is created so it knows what the last move was. The
prev_move_pos is something that already happened in the tree, so we
never want to change it once it is set.